doctor-app/requirements.txt
streamlit>=1.28.0
pydantic_ai>=0.1.2
openai>=1.0.0
python-dotenv>=1.0.0
mcp>=1.6.0
anyio>=4.5
click>=8.1.0
httpx>=0.27
uvicorn>=0.23.1
starlette>=0.27
